refactor(logging): move multi node logging to a separate file
